% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_indicators}
\alias{add_indicators}
\title{Adicionar los indicadores como total casos,
total muestras y positividad}
\usage{
add_indicators(
  data_grouped,
  report_data = NULL,
  col_name = NULL,
  total_cases = TRUE,
  total_samples = FALSE,
  positivity = FALSE,
  remove_nan = TRUE,
  join = TRUE
)
}
\description{
Adicionar los indicadores como total casos,
total muestras y positividad
}
