% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_missing_weeks}
\alias{add_missing_weeks}
\title{Añadir las semanas epidemiológicas faltantes}
\usage{
add_missing_weeks(
  dataset,
  col_epiweek,
  col_casos = "casos",
  col_total = "total_casos",
  col_porcentaje = "porcentaje"
)
}
\description{
Añadir las semanas epidemiológicas faltantes
}
