% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_filmarray_age}
\alias{clean_filmarray_age}
\title{Estandarizar las edades de los datos de la base de datos de la
Fundación Cardio Infantil}
\usage{
clean_filmarray_age(filmarray_data, col_age = "edad", is_panel = FALSE)
}
\description{
Estandarizar las edades de los datos de la base de datos de la
Fundación Cardio Infantil
}
