% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{combine_epi_periods}
\alias{combine_epi_periods}
\title{Combinar la informacion historica de los virus
con la calculada del año actual}
\usage{
combine_epi_periods(
  data_historic_epi_periods,
  data_current_year_epi_periods,
  epi_period,
  col_name = "periodo_epidemiologico",
  year
)
}
\description{
Combinar la informacion historica de los virus
con la calculada del año actual
}
