% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complete_epiweeks}
\alias{complete_epiweeks}
\title{Completar las semana epidemiológicas de un data.frame}
\usage{
complete_epiweeks(dataset, col_epiweek)
}
\description{
Completar las semana epidemiológicas de un data.frame
}
