% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{fill_down_column}
\alias{fill_down_column}
\title{Rellenar valores faltantes en una columna}
\usage{
fill_down_column(dataset, column_name)
}
\arguments{
\item{dataset}{Un dataset que contiene la columna a rellenar.}

\item{column_name}{Nombre de la columna a rellenar (como variable sin comillas).}
}
\value{
Un dataset con los valores de la columna completados.
}
\description{
Usa `tidyr::fill()` para rellenar los valores faltantes en la columna especificada,
propagando los valores hacia abajo.
}
