% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_all_tables}
\alias{get_all_tables}
\title{Extraer todas las tablas de una hoja de Excel}
\usage{
get_all_tables(file_name)
}
\arguments{
\item{file_name}{Ruta del archivo de Excel.}
}
\value{
Lista de dataframes, donde cada elemento representa una tabla identificada dentro de la hoja.
}
\description{
Detecta y extrae múltiples tablas dentro de una hoja de Excel, identificando 
separaciones mediante filas y columnas en blanco.
}
