% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_axis_config_periodo_epidemiologico}
\alias{get_axis_config_periodo_epidemiologico}
\title{Obtener configuración de los ejes para el periodo epidemiológico}
\usage{
get_axis_config_periodo_epidemiologico()
}
\value{
Lista con la configuración de los ejes:
- `y_axis1_max_value`: Valor máximo del primer eje Y.
- `y_axis2_max_value`: Valor máximo del segundo eje Y.
- `scaling_factor`: Factor de escala entre los dos ejes.
- `bar_width`: Ancho de las barras.
- `line_width`: Ancho de la línea.
}
\description{
Devuelve una lista con los valores máximos de los ejes Y, el factor de escala 
y los parámetros de ancho de barra y línea.
}
