% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_acum_proportion}
\alias{get_cases_acum_proportion}
\title{Obtener la proporción acumulada de los virus por semana
epidemiológica}
\usage{
get_cases_acum_proportion(dist_epiweek_viruses, epiweek = "all")
}
\description{
Obtener la proporción acumulada de los virus por semana
epidemiológica
}
