% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_influenza}
\alias{get_cases_influenza}
\title{Obtener la distribución de casos de Influenza y sus subtipos}
\usage{
get_cases_influenza(filmarray_data, other_virs_data, epiweek = "all")
}
\description{
Obtener la distribución de casos de Influenza y sus subtipos
}
