% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_color_periodo_epidemiologico}
\alias{get_color_periodo_epidemiologico}
\title{Obtener configuración de colores para el periodo epidemiológico}
\usage{
get_color_periodo_epidemiologico()
}
\value{
Lista con los colores definidos
}
\description{
Devuelve una lista con los colores asignados a diferentes elementos del gráfico, 
incluyendo líneas, títulos de ejes, líneas verticales y tipos de virus.
}
