% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_consolidated_viruses}
\alias{get_consolidated_viruses}
\title{Obtener los valores consolidados de los virus que en los datos se
encuentran en múltiples columnas, como por ejemplo Parainfluenza y Otros
coronavirus}
\usage{
get_consolidated_viruses(
  dataset,
  col_name = "grupo_edad",
  percentage = TRUE,
  total_cases = TRUE
)
}
\description{
Obtener los valores consolidados de los virus que en los datos se
encuentran en múltiples columnas, como por ejemplo Parainfluenza y Otros
coronavirus
}
