% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_distribution_test}
\alias{get_distribution_test}
\title{Obtener la distribución de casos por tipo de prueba}
\usage{
get_distribution_test(
  report_data,
  epiweek = "all",
  include_sars = FALSE,
  test = NULL,
  col_epiweek = NULL
)
}
\description{
Obtener la distribución de casos por tipo de prueba
}
