% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_epi_times_current_year}
\alias{get_epi_times_current_year}
\title{Obtener los casos, total muestras y positividad del
año actual}
\usage{
get_epi_times_current_year(
  data_epiweek,
  year,
  col_name = "se",
  period_epi = NULL,
  remove_cols = c("h1n1", "otros_coronavirus")
)
}
\description{
Obtener los casos, total muestras y positividad del
año actual
}
