% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rows_valid_str_detect}
\alias{get_rows_valid_str_detect}
\title{Obtener las columnas de un data.frame según los resultados de
str_detect y una validación de NANs}
\usage{
get_rows_valid_str_detect(cases_virus, col_name, values, is_not = FALSE)
}
\description{
Obtener las columnas de un data.frame según los resultados de
str_detect y una validación de NANs
}
