% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_text_labels_periodo_epidemiologico}
\alias{get_text_labels_periodo_epidemiologico}
\title{Generar etiquetas de texto para los periodos epidemiológicos}
\usage{
get_text_labels_periodo_epidemiologico(dataset_epi_time)
}
\arguments{
\item{dataset_epiTime}{Dataset con una columna `ano` que contiene los años.}
}
\value{
Vector de texto con etiquetas para cada año.
}
\description{
Extrae los años únicos del dataset y genera etiquetas en formato "AÑO <año>".
}
