% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_top_virues_age}
\alias{get_top_virues_age}
\title{Obtener el top de virus por grupo de edad}
\usage{
get_top_virues_age(
  dataset,
  col_age = "grupo_edad",
  age_groups = c("< 2 A<U+00D1>OS", "2 A 4 A<U+00D1>OS")
)
}
\description{
Obtener el top de virus por grupo de edad
}
