% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_total_samples}
\alias{get_total_samples}
\title{Obtener el total de muestras}
\usage{
get_total_samples(
  data_grouped = NULL,
  report_data = NULL,
  col_name = NULL,
  join = TRUE
)
}
\description{
Obtener el total de muestras
}
