% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_historic}
\alias{import_data_historic}
\title{Importar los datos historicos de los virus respiratorios}
\usage{
import_data_historic(file_name)
}
\description{
Importar los datos historicos de los virus respiratorios
}
