% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_viral_circulation}
\alias{import_data_viral_circulation}
\title{Importar los datos de la circulación viral}
\usage{
import_data_viral_circulation(
  report_data = NULL,
  header = FALSE,
  skip = NULL,
  col_names = FALSE,
  sheet = NULL,
  dataset_name = NULL
)
}
\description{
Importar los datos de la circulación viral
}
