% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_distribution_epiweek}
\alias{plot_distribution_epiweek}
\title{Graficar la distribución de casos por semana epidemiológica}
\usage{
plot_distribution_epiweek(
  report_data,
  var_x = "semanaepidemiologicavegeneral",
  var_y = "casos",
  var_fill = "etiqueta",
  var_total = "total_casos",
  var_positives = "porcentaje",
  influenza = FALSE,
  positives = NULL
)
}
\description{
Graficar la distribución de casos por semana epidemiológica
}
