% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_row}
\alias{rename_row}
\title{Renombrar los valores de una o más filas de un data.frame}
\usage{
rename_row(dataset, col_name, original_names, new_name)
}
\description{
Renombrar los valores de una o más filas de un data.frame
}
