% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{save_vals_cases_epiweeks}
\alias{save_vals_cases_epiweeks}
\title{Crear y guardar el archivo de virus respiratorios por semana
epidemiológica del año actual del reporte}
\usage{
save_vals_cases_epiweeks(viruses_epiweeks, year)
}
\description{
Crear y guardar el archivo de virus respiratorios por semana
epidemiológica del año actual del reporte
}
